fix tests

+++ src/Core/Video/UseCase/UpdateUseCase.php
@@ -5,6 +5,7 @@ namespace Core\Video\UseCase;
 use Core\Video\Builder\VideoUpdateBuilder;
 use Core\Video\Domain\Entity\Video;
 use Core\Video\Interfaces\VideoBuilderInterface;
+use Costa\DomainPackage\UseCase\Exception\NotFoundException;
 use Costa\DomainPackage\UseCase\Exception\UseCaseException;
 use Throwable;
 
@@ -17,30 +18,34 @@ class UpdateUseCase extends BaseUseCase
 
     public function execute(DTO\Update\Input $input): DTO\Update\Output
     {
-        try {
-            $this->builder->createEntity($input);
-            $this->createEntity($input);
-
-            if ($this->repository->update($this->builder->getEntity())) {
-                $filesUploads = $this->storageAllFiles($input);
-                $this->repository->updateMedia($this->builder->getEntity());
-                $this->eventManager->dispatch($this->builder->getEntity());
-                $this->transaction->commit();
-
-                return $this->output($this->builder->getEntity());
-            }
-        } catch (Throwable $e) {
-            $this->transaction->rollback();
-            if (isset($filesUploads)) {
-                foreach ($filesUploads as $file) {
-                    $this->storage->delete($file);
+        if ($obj = $this->repository->findById($input->id)) {
+            try {
+                $this->builder->createEntity($obj);
+                $this->verifyCategories($input);
+                $this->verifyGenres($input);
+                $this->verifyCastMembers($input);
+                if ($this->repository->update($this->builder->getEntity())) {
+                    $filesUploads = $this->storageAllFiles($input);
+                    $this->repository->updateMedia($this->builder->getEntity());
+                    $this->eventManager->dispatch($this->builder->getEntity());
+                    $this->transaction->commit();
+                    return $this->output($this->builder->getEntity());
                 }
+
+                throw new UseCaseException(self::class);
+                
+            } catch (Throwable $e) {
+                $this->transaction->rollback();
+                if (isset($filesUploads)) {
+                    foreach ($filesUploads as $file) {
+                        $this->storage->delete($file);
+                    }
+                }
+
+                throw $e;
             }
-
-            throw $e;
         }
-
-        throw new UseCaseException(self::class);
+        throw new NotFoundException($input->id);
     }
 
     protected function output(Video $entity)
